add confident score

+++ DataLoader.py
@@ -2,6 +2,7 @@ from torch.utils.data import DataLoader
 import torch
 import torch.nn.utils.rnn as rnn_utils
 from torch_geometric.loader import DataLoader as GCNDataLoader
+from train_val import dtype
 
 
 def rnn_collate_fn(data):
@@ -30,8 +31,8 @@ def stgcn_collate_fn(data):
         int_label.append(d[1][0])
         att_label.append(d[1][1])
         act_label.append(d[1][2])
-    return input, (torch.Tensor(int_label).to(torch.float16), torch.Tensor(att_label).to(torch.float16),
-                   torch.Tensor(act_label).to(torch.float16))
+    return input, (
+    torch.Tensor(int_label).to(dtype), torch.Tensor(att_label).to(dtype), torch.Tensor(act_label).to(dtype))
 
 
 class JPLDataLoader(DataLoader):
